Fix crash in skia graphics context teardown after gpu context loss
GrContextForWebGraphicsContext3D::OnLostContext() was
calling GrContext::contextLost() which has the wrong
behavior: it assumes that the gpu context is still valid
but that all of its resources have been lost, so it
attempts to reinitialize resources that are used internally
by skia. The correct action is to call
GrContext::contextDestroyed(),  which signals that the gpu
context is no longer usable.

BUG=361555

Review URL: https://codereview.chromium.org/233583002