Revert "Revert 262050 "Implement a source for remote video tracks.""
This reverts commit c6d743ef45ab10d0e014dd07b81de10a9aaa3e0c.

The underlaying problem is fixed in
https://codereview.chromium.org/232803002/

Original description:
Implement a source for remote video tracks.
MediaStreamRemoteVideoSource implements the MediaStreamVideoSource interface
for video tracks received on a PeerConnection. The purpose of the class is
to make sure there is no difference between a video track where the source is
a local source and a video track where the source is a remote video track.

Verified on ChromeOs with hw decode support.

TBR=ronghuawu

BUG=334243

Review URL: https://codereview.chromium.org/231983006