Store the language code for the address in autofill profile.
After BuildComponents() returns the language code that should be used to
format the address, Chrome should store and sync this language code
along with the autofill profile. The language code should not be
communicated to the autofill server, because it's not the type of data
that will be sent to an input field on a webpage.

BUG=354955

Review URL: https://codereview.chromium.org/212873003