Mac: Fix throbber not spinning after dragging tabs.
We want the throbber to spin when it's attached to a window, and that window is
not miniaturized.
We were calling -updateAnimation: from -viewWillMoveToWindow where [self window]
is still the old window. We want to update the animation based on the new
window, so the fix is to call it from -viewDidMoveToWindow.

BUG=361969

Review URL: https://codereview.chromium.org/235583002