cc: Handle retroactive BeginFrames in the Scheduler.
This will be the first step allowing us to pull logic out
of the OutputSurface.

Initially, the retroactive logic will exist in both the
OutputSurface and the scheduler. Then, as all logic that
depends on the retroactive logic has moved to the
scheduler, I will remove the retroactive logic in
OutputSurface.

There are now 4 BeginFrames in the scheduler:
1) BeginFrame: from the OutputSurface.
2) BeginRetroFrame: a deferred BeginFrame.
3) BeginImplFrame: starts a compositor frame.
4) BeginMainFrame: starts a main frame.

BUG=358409

Review URL: https://codereview.chromium.org/218633010