Get horizontal wheel scrolling working again with logitech drivers.
Some Logitech drivers send WM_MOUSEHWHEEL messages to the parent window intermingled with WM_MOUSEWHEEL
messages to the child if there is one. This causes horizontal scrolling to be treated as vertical scroll.

Hack is to remember the last time a WM_MOUSEHWHEEL message is received and treat WM_MOUSEWHEEL messages
which occur in this window as WM_MOUSEHWHEEL messages.

BUG=332797
R=sky

Review URL: https://codereview.chromium.org/236183004