Shorten the slave build directory name for all nacl builders.
We've hit the character limit for windows in a few cases recently.
As our builds are always clobbers, this sharing should be ok (we do it on the
trybots already).

Shortening only the slave build-dir name as on the master they are not allowed
to collide.

BUG=None
TEST=None
R=agable@chromium.org,dyen@chromium.org,noelallen@chromium.org

Review URL: https://codereview.chromium.org/237183004