Telemetry: Reduce the default number of pixels covered by a pinch gesture.
The current value for the number of pixels covered by a pinch is very large (4000).
On most pages this means that we keep pinching for a while after we've reached
the maximum zoom level. When we lift the pointers, the scrollbar fades out, thus
causing a frame to get rendered. The gap between reaching the maximum zoom level
and the scrollbar fading out is considered as jank by Telemetry. Reducing the
default number of pixels covered by a pinch to a more sensible value (500) seems
to mitigate this problem.

BUG=

Review URL: https://codereview.chromium.org/238023005