Show error reports from ServiceWorker on serviceworker-internals.
In this cl we introduce ServiceWorkerContextObserver and make ServiceWorkerInternalsUI inherit from it.

In the constructor of ServiceWorkerContextWrapper we create ObserverListThreadSafe<ServiceWorkerContextObserver> and pass it to ServiceWorkerContextCore.
The methods of ServiceWorkerContextObserver will be called via ObserverListThreadSafe::Notify() called from ServiceWorkerContextCore on IO thread.
ServiceWorkerInternalsUI is registered to ObserverListThreadSafe in ServiceWorkerContextWrapper::AddObserver() on UI thread so these methods will be on UI thread.


BUG=359517

Review URL: https://codereview.chromium.org/236153004