Break a single histogram into two histograms
This histogram records a networking result that if positive indicates a
byte count, and if negative indicates an error code. Since histograms
can't handle negative values, this splits that up into two: one for the
byte count, and one that will use an existing enum for the network
error code.

This was based off of a suggestion in codereview.chromium.org/236023005

BUG=360909

Review URL: https://codereview.chromium.org/235893017