Fix artifact appearing on dragging an item out of an app list folder [Win]
This CL fixes an issue where a thin line would appear and animate when an
item was dragged out of app list folder. The view is now properly set visible
which replaces the artifact with the full app icon. This is consistent
with the behavior on ChromeOS.

BUG=363007

Review URL: https://codereview.chromium.org/236203021