Fix Window size issues with high dpi enabled on Windows.
This regressed with oshima's change https://codereview.chromium.org/201573015/ which introduced AshWindowTreeHost
and did a bunch of refactoring and reorganization in the WindowTreeHost and associated classes.

The bug was that we were setting pixel bounds to aura::Window which basically causes the window to size abnormally. This occurred because we were using WindowTreeHost::GetRootTransform() in the WindowTreeHost::UpdateRootWindowSize function.
The root transform also scales the window back to the current scale factor which converts the window bounds to 
pixel.

We need to mimic the old code which used an empty transformer, which in this case is the default transformer 
in the window layer.

BUG=363289,350327
R=sky

Review URL: https://codereview.chromium.org/237673012