Allow mouse events to propagate in WindowOverview::OnMouseEvent().
Use event->SetHandled() instead of StopPropagating so that MouseCursorEventFilter
can still see the event but it is prevented from being sent to the renderer.

When cursor is inside a target window in overview mode, we still want to propagate
MOUSE_MOVED events so that they can be received by MouseCursorEventFilter to update
the cursor window.

BUG=chrome-os-partner:28034
TEST=trybot and manual verification

Review URL: https://codereview.chromium.org/238253009