Use invalid value for GetAccessPointName for CHROME_IOS_RESERVED
The team that uses the RLZ access point does not want to reserve two
access points for Chrome on iOS since only one is currently used but
the enumeration rlz_lib::AccessPoint cannot be changed.

Mark the CHROME_IOS_RESERVED access point enumeration value as not
supported in IsAccessPointSupported (as we should never reference an
access point with enumeration value until the value is reused). This
fixes the code that iterates over all values in that enumeration in
FinancialPing::FormRequest when no events have been recorded.

Return an invalid value in GetAccessPointName if called with the
enumeration value rlz_lib::AccessPoint instead of returning NULL to
avoid a crash on Chrome OS (constructing a std::string from NULL crash
in strlen on Chrome OS).

BUG=360686,363700

Review URL: https://codereview.chromium.org/239053004