Add a DistilledContentStore (and an in-memory impl)
This is the interface to whatever is used for storing and fetching
distilled content. In the future, this should be provided by sync
attachments. For now, we need something.

The interface is a very simple pair of functions, SaveContent and
LoadContent, both asynchronous.

There is a simple in-memory implementation that just stores saved
content in a map.

The DomDistillerService creates the store and uses a save callback on
the TaskTracker to save content to it. The TaskTracker receives a
pointer to the store and uses it when StartBlobFetcher is called.

BUG=288015

Review URL: https://codereview.chromium.org/189833002