Add generate_jni template
This adds a template for using the jni_generator. Creates the
base_jni_headers target and updates the base target to depend on it and
actually use the sources that required it.

Adds //third_party/android_tools:cpu_features required by one of the
base files that also required jni.

Disable a bunch of third_party targets for Android builds.

BUG=359249

Review URL: https://codereview.chromium.org/237973002