adjust retry logic in chromium_gn recipe
Previously we would attempt to check out the code, apply the patch, and
run the hooks, and only then check for error codes and nuke the checkout
is something failed. If the checkout step fails, there's no point in
applying the patch or running the hooks, so this patch now nukes after
the first checkout.

We could theoretically nuke *again* if the checkout works but applying
the patch or running the hooks fails, but that seems like a pretty rare
occurrence, so we'll leave it be for now.

If this turns out to be a real source of flaky try jobs failures, we can
add it back in.

R=phajdan.jr@chromium.org,
BUG=362226

Review URL: https://codereview.chromium.org/228473004