Revert 244720 "When installations are in progress only delay cle..."
Unfortunately this had the side effect of exposing a race conditions that
could result in actually valid extensions being garbage collected. Since
the real fix for that is too big to backport, reverting back to not
garbage collecting at all in certain situations is the lesser of the two
problems.

BUG=342271,365469

> When installations are in progress only delay cleaning up of Temp directory, don't delay cleaning up of old extension versions.
> 
> This makes the effect of chrome never finishing installing pending extension less severe since most data will still be cleaned up.
> 
> BUG=329943
> 
> Review URL: https://codereview.chromium.org/131153003

TBR=mek@chromium.org

Review URL: https://codereview.chromium.org/245843003