Add missing library dependencies to libc++abi.so.
Fixes a link error in mksnapshot. Previously, only libc++.so had those
dependencies - but libc++abi.so needs them as well. Usually this isn't a
problem because executables tend to depend on both libc++ and libc++abi, but in
at least one case the libc++.so dependency was dropped due to -Wl,--as-needed,
leaving libc++abi.so with unresolved symbols.

BUG=365612
TBR=glider@chromium.org
TEST=compile
NOTRY=true

Review URL: https://codereview.chromium.org/246763004