cc: Fix missing compositor invalidations during resize
In practice, cc is responsible for implicit invalidations on layers,
such as the original layer being entirely invalid or new areas being
exposed during resize.

In this case, these invalidations were being clipped to the live tiles
rect, which itself is clipped to the bounds of the tiling.  Therefore,
when the bounds of a layer grew, the new invalidation intersected with
the old bounds was empty.

The intersection of invalidations with the live tiles rect is an
optimization, to prevent needless work trying to look up invalidations
on parts of the layer that don't matter.

To fix the bug, rather than intersecting with the live tiles rect,
expand the live tiles rect to tile boundaries.  This lets implicit
invalidation due to resize within a tile not skip rerasterizing
that tile.

BUG=357120

Review URL: https://codereview.chromium.org/240593004