Update platform_app.js to work even if window.history is not replaceable
platform_app.js was using "history = {};" to disable the History API. However,
Blink is about to be updated so that window.history is no longer replaceable
(as per the specification, and to align with other browsers). This CL updates
platform_app.js so that history is still disabled, even if window.history is
not replaceable.

BUG=365355
TEST=PlatformAppBrowserTest.Restrictions

Review URL: https://codereview.chromium.org/245443003