sync: Sticky scroll-to-bottom on about:sync
Updates the sync protocol events UI to keep scrolling down if new events
are added while the scroll bar is already at the bottom.  If the logs
were not scrolled to the bottom, then the scrollTop is left unmodified.

Think of it as "gnome-terminal" style scrolling.

Also fixes a few comments in about.js.

BUG=362820

Review URL: https://codereview.chromium.org/246643011