ServiceWorker should be able to send messages to browser (not only for reply)
Currently ServiceWorker (in an EmbeddedWorker) can only send a 'reply'
message to the browser, as ServiceWorkerScriptContext::SendMessageToBrowser()
always takes |request_id| and the message is handled based on the |request_id|
in the receiver side, but in order to implement more features (e.g. .clients)
it needs to be able to initiate the communication.

This patch:
- Splits SendMessageToBrowser message into two: MessageToBrowser and ReplyToBrowser
- Renames SendMessageTo* message names to MessageTo* or ReplyTo* (cleanup, to avoid confusion)
- Renames OnSendMessageTo* message handlers to OnMessageTo* or OnReplyTo* (ditto)
- Renames EmbeddedWorkerInstance::Observer to ::Listener to make it clear what it does

BUG=363965
TEST=ServiceWorkerVersionTest.ReceiveMessageFromWorker
R=falken@chromium.org, jam@chromium.org, tsepez@chromium.org

Review URL: https://codereview.chromium.org/240243003