Rename InvalidationAuthProvider to IdentityProvider
The InvalidationAuthProvider is useful outside the area of invalidation:
It provides an abstract interface for accessing the logged-in GAIA account
and receiving notifications about login/logout events. This CL renames
the InvalidationAuthProvider to IdentityProvider and places it in the
gaia_apis component.

This change is a prerequisite for CL 225403021, which will use the
IdentityProvider in GCM.

BUG=362083
TEST=Updated tests

TBR=atwilson (for chrome/browser/sync/test/integration/sync_test.cc)

Review URL: https://codereview.chromium.org/235273002