Currently, we try to set virtual keyboard window height both in native code and in VK's javascript code through window.resizeTo It sometimes creates conflict and cause strange animation issues.
In this CL, the window height logic in native side is removed.
The default window height is set to 0 and we expect javascript
will set the window to correct height.
If window is resized from a non zero height to another non
zeor height, we current dont do any animation.

BUG=363622

Review URL: https://codereview.chromium.org/240443006