android: Attach ContentViewCore to RenderWidgetHostViewAndroid when swapping RenderViewHost
Previously, we unattached ContentViewCore from RenderWidgetHostViewAndroid when
swapping out the RenderHostView.  We also need to attach the ContentViewCore
on the swapped-in RenderWidgetHostViewAndroid.  We signal a resize to the
renderer after attaching a valid ContentViewCore.

This patch also made sure that WebContentsView::GetContainerBounds() returns the
correct size for the container and not defer to RenderWidgetHostViewAndroid.

Removed WasResized() from RWHVAndroid::SetSize() as (a) it is redundant for size
changes triggered from the content layer and (b) it will clobber with the wrong
value if the RWHV hasn't been attached and properly sized yet.

BUG=358252

Review URL: https://codereview.chromium.org/233803002