Mark touchcancel events as uncancelable
The touch events spec says that all touch events are cancelable except for
touchcancel, and this is exposed to JS via the cancelable property.
Previously we were marking all touch events as cancelable, but as of
https://src.chromium.org/viewvc/blink?revision=172329&view=revision the
decision is up to the browser (since only it can decide if it will wait for
an ACK).  This change ensures all touchcancel events are marked as
uncancelable, and all other touch events remain cancelable. 

There are a lot of different places that create WebTouchEvents, and we
don't want some central location (like the InputRouter) to be modifying
these events.  So I've added a WebTouchEventTraits::ResetType function to
reset properties of a WebTouchEvent, applying appropriate default policy
like cancelable state.

Then we can DCHECK at the central location that an a event is cancelable
if and only if we intend to wait for it's ACK.

I've updated a handful of unit tests to verify that the cancelable bit 
is set correctly in various scenarios.

This also extends EventSender to allow the cancelable bit to be set so that
we can add a blink LayoutTest, and removes an unused
SyntheticWebTouchEventBuilder.

BUG=365681

Review URL: https://codereview.chromium.org/247433003