Prepare browser_tests for data: urls being async
RenderViewTest::LoadHTML assumed that issuing a load
of a data url, posting a task to quit the message loop
and then pumping the message loop until it quits was
sufficient to end up with a fully parsed page.

That very soon won't be true once we send data: urls
to the threaded parser, as the parser thread posts
back to the main thread and may post tasks after the
quit message.
https://codereview.chromium.org/210253003/

Instead I've added a FrameLoadWaiter class which
implements RenderViewObserver to listen for the
load event for the frame and only post the quit
task once the load event had come in.

BUG=308321

Review URL: https://codereview.chromium.org/257723002