Multiprofile: switch to newly added user to the session
This CL remember pending user switch and do it as soon as profile
ready. From UI perspective this look very smooth because profile loading
is actually fast so previous desktop is not shown. But if it will be
a problem on slow device we may need to stay on add user screen a bit longer.

BUG=333879
TEST=manual
R=nkostylev@chromium.org

Review URL: https://codereview.chromium.org/246253009