Don't throw an exception if onfullscreenchanged fires after the plugin is removed.
We explicitly cancel full screen mode on disconnection, which was resulting in
an exception in the event handler, preventing the full-screen class from being
removed from the <html> node. The left the plugin container full-screen on top
of the Disconnected dialog, making it unclickable.

BUG=365920
NOTRY=true

Review URL: https://codereview.chromium.org/258863003