Reland "Implement async touchmove dispatch during scroll"
The patch landed originally in r266470, but was reverted in r266579 due to an
uninitialized variable.  The fix was reviewed in crrev.com/253813003.

This implements a touch dispatch model in which touchmove sending is throttled
while a scroll sequence is active and being consumed. Such touchmove's are
marked with |cancelable = false|, indicating to any potential consumers that the
event cannot be prevented from triggering a platform gesture. Throttling limits
the touchmove sending rate during scrolling to 1 event per 200ms.

BUG=346693,367692
NOTRY=true
TBR=jochen@chromium.org

Review URL: https://codereview.chromium.org/254803012